Fix issue where watchdog being triggered on core 0
This issue seemed to stem from the size of the stack for the task. It was set to 10000, and I set it down to 1000. This cause guru meditation errors with stack overflow, so I bumped it back up to 2000 and it seems to be working fine now. Also had a concern where analog reads appeared to be taking up to 20ms, but this may have been due to the Serial.prints taking up time. Commenting out all of the printData print messages made the analog reads instantaneous (at least from the perspective of printing the before and after millis() )