Add "Ignore Readings" Code
This code is meant to ignore any calculations that may happen after a long task, such as printing to the serial monitor or sending CAN data. Since we are not actively monitoring the analog reads during this tie, there is the chance that we could miss a revolution which would in turn mess up calculations.