Temporary Fix for data not updating
I set the ignoreReadings to false because my hand can only go so fast. However, this presents another issue for the car. If we are only getting something like three revolutions per second, we will never do the math to update because we are sending CAN data 10 times per second. Need to find a solution for this