Remove ignore reading code
Although the serial printing takes ~7000us, it is not worth having ignore reading code. If the readings are too slow (fewer readings per second than instances of printing to serial monitor), then we will never do calculations. The serial printing will not be in place in the final version, so it will have no effect on final accuracy. The CAN sending is nearly instantaneous compared to serial printing, so we do not have to worry about that too much