Add version for READING READING MATH instead of READING MATH READING
This is for if we have 5+ alternating dark/light spots on the CVT. At max RPM, it would be too slow to calculate the RPM every time we get a new reading. We can get away with just waiting for two consecutive readings, then doing the math. Even if we miss some while doing the math, we always use two readings with no interruptions so it should still be accurate. However, a new variable is needed that divides the RPM based on how many targets there are. For example, if we have 8 targets, then we need to divide the calculated RPM by 8