Several updates
I think the status bits / RTRs may have been causing issues, so those are removed.

In the case that extremely long CAN messages (like sending 250000 RPM) are causing the issues on startup (when the CVT sensors have not been calibrated), I added a statement that first checks if the CVT has calibrated itself. If it has not, it jut sets the RPM value to zero