See description
Untested code, adds some features for better accuracy when sampling. Now, ADC takes five samples of the IR pin and averages before using primaryValue and secondaryValue in calculations.

abs was added to the threshold checker. This probably won't make a difference, but makes the code slightly more robust against weird errors and also easier to understand/read

ignoreNextReading was added to the secondary code. Especially with the new averaging, sending data to the CAN bus and delaying 1ms for the watchdog could be enough to miss or just get the edge of a sensing target. Better safe than sorry, missing one reading is not crucial to data interpretation but having an incorrect one is annoying to fix in post.